Switch to using injectable to setup route for stopping current port forward
