+++ src/main/routes/port-forward-route.ts
@@ -4,7 +4,6 @@
  */
 
 import type { LensApiRequest } from "../router";
-import logger from "../logger";
 import { respondJson } from "../utils/http-responses";
 import { PortForward } from "./port-forward/port-forward";
 
@@ -21,28 +20,5 @@ export class PortForwardRoute {
     });
 
     respondJson(response, { port: portForward?.forwardPort ?? null });
-  }
-
-  static async routeCurrentPortForwardStop(request: LensApiRequest) {
-    const { params, query, response, cluster } = request;
-    const { namespace, resourceType, resourceName } = params;
-    const port = Number(query.get("port"));
-    const forwardPort = Number(query.get("forwardPort"));
-
-    const portForward = PortForward.getPortforward({
-      clusterId: cluster.id, kind: resourceType, name: resourceName,
-      namespace, port, forwardPort,
-    });
-
-    try {
-      await portForward.stop();
-      respondJson(response, { status: true });
-    } catch (error) {
-      logger.error("[PORT-FORWARD-ROUTE]: error stopping a port-forward", { namespace, port, forwardPort, resourceType, resourceName });
-
-      return respondJson(response, {
-        message: `error stopping a forward port ${port}`,
-      }, 400);
-    }
   }
 }
